feat: Handle permissions for hall Rates

+++ resources/views/components/pages/hall-rates/table/⚡table.blade.php
@@ -37,6 +37,8 @@ new class extends Component {
     #[On('delete-hall-rate')]
     public function deleteHallRate(int $id): void
     {
+        $this->authorize('delete', HallRate::class);
+
         $hallRate = HallRate::findOrFail($id);
 
         $hallRate->delete();
@@ -60,11 +62,13 @@ new class extends Component {
             :placeholder="__('forms.search')"
         />
 
-        <button type="button"
-                wire:click="$dispatch('open-modal', 'openCreateModal')"
-                class="flex flex-row items-center gap-2.5 px-4 py-3 border border-brown bg-brown text-white group rounded-sm hover:bg-white hover:text-brown trans-all justify-center md:col-start-3 md:col-span-2 md:justify-self-end">
-            {{ __('pages/hall.hall-rates.add-hall-rate') }}
-        </button>
+        @can('hallRates.create')
+            <button type="button"
+                    wire:click="$dispatch('open-modal', 'openCreateModal')"
+                    class="flex flex-row items-center gap-2.5 px-4 py-3 border border-brown bg-brown text-white group rounded-sm hover:bg-white hover:text-brown trans-all justify-center md:col-start-3 md:col-span-2 md:justify-self-end">
+                {{ __('pages/hall.hall-rates.add-hall-rate') }}
+            </button>
+        @endcan
     </div>
 
     <x-general.selected-column
@@ -72,7 +76,7 @@ new class extends Component {
         :options="[
             'delete' => true,
             ]"
-        :delete-permission="null"
+        delete-permission="hallRates.delete"
     />
 
     @if($this->getHallRates->isNotEmpty())
